ims_registrar_pcscf: sec-agree support for multiple security mechanisms
A REGISTER message with multiple security mechanisms in Security-Client
header can be received. The current implementation of the parser doesn't
support this.

This patch adds support for mutiple mechanisms.

!!! WIP warning: Effectively this patch saves only the last security
mechanism in the contact. !!!

Additionaly it parses only Security-Client parameters from the REGISTER
in order to avoid saving bad IPSec parameters.